Added an abstraction to publish MQTT messages (#235)
